Create Basic Game Logic
- Added Card.jsx: Card component displays the image data given to it &
has a onChange callback function to pass the id.
as props.
- Update App.jsx: Add score state, and imgData now maintains the initial
order of fetched images in `initial`, while `current` is used during
gameplay.
Add randomize function that suffles the array elements.
Add handleClick function that takes id, and updates the imgData and
score states.
If the user clicks the unique card, the score increases, and cards gets
suffled randomly.
If the user clicks the same card twice, the game resets, saving the
highscore.
Also returns dom elements to display the score values.